Make the postgis extension generator into a CLI command.
