+++ edb/tools/postgis.template.edgeql
